Update PFP and GifSystem cogs
