+++ cogs/GifSystem.py
@@ -1,5 +1,6 @@
-from discord.ext import commands
 import discord
+from discord import app_commands
+from discord.ext import commands
 import json
 import aiohttp
 import random
@@ -10,11 +11,14 @@ class GifSystem(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    @commands.command(name='giphy', aliases=['g'], pass_context=True)
-    async def giphy(self, ctx, *, search=None):
+    @app_commands.command(
+        name="gif",
+        description="Get a random gif from Giphy")
+    @app_commands.describe(
+        search="The search term to query Giphy for")
+    async def gif(self, interaction: discord.Interaction, *, search: str = None) -> None:
         """
         Get a random gif from giphy
-        :param ctx: context object
         :param search: search query
         """
         session = aiohttp.ClientSession()
@@ -31,10 +35,11 @@ class GifSystem(commands.Cog):
             embed.set_image(url=data['data'][gif_choice]['images']['original']['url'])
 
         await session.close()
-        await ctx.message.delete()
-        await ctx.send(embed=embed)
+        await interaction.response.send_message(embed=embed)
 
 
-def setup(bot):
-    bot.add_cog(GifSystem(bot))
-    print('GifSystem is loaded')
+async def setup(bot: commands.Bot):
+    await bot.add_cog(
+        GifSystem(bot),
+        guilds=[discord.Object(id=980975086154682378)]
+    )
